+++ src/preload.cjs
@@ -10,5 +10,6 @@ contextBridge.exposeInMainWorld('taxLedger', {
   exportCsv: (store, year) => ipcRenderer.invoke('store:export-csv', store, year),
   exportPdf: (store, year) => ipcRenderer.invoke('report:export-pdf', store, year),
   openFolder: (filePath) => ipcRenderer.invoke('app:open-folder', filePath),
-  getVersion: () => ipcRenderer.invoke('app:version')
+  getVersion: () => ipcRenderer.invoke('app:version'),
+  onMenuAction: (callback) => ipcRenderer.on('menu:action', (_event, action) => callback(action))
 });